Refactor: Adding vitest to Avatar component (#2622)
